Add a 'tutorial' folder to the distribution, with an initial example.
It would be nice to gather neat Dafny examples there; each new feature could
have its own small file that demoes it, and we could also have examples that
showcase stuff that we think is impressive.

I'm adding this as a test folder, because it's important to check that these
cool examples don't break, but the focus probably shouldn't be on exhaustively
testing the features being demoed.