Add Dafny reference manual
This version is still a draft, but is mostly complete and
about half reviewed. The manual is written using
Madoko. The sources are in the Docs/DafnyRef directory.
The processed sources are available in the
Docs/DafnyRef/out directory in the form of a single
HTML page or as a PDF.