Patch Dafny to use the system Z3
In previous versions, it was sufficient to just pass /z3exe:/usr/bin/z3
in the dafny launcher script, but now, Dafny looks for its vendored
Z3 (which we don’t install) at runtime and prints a warning if it’s
gone.  Patch that behaviour away, hardcoding /usr/bin/z3 as the Z3
binary location.